Added trigger to allow actions before duplicate
Added jQuery trigger to allow actions before duplication of an entry. Use case: The Neo fields do not get duplicated if the hidden 'input.modified' field is set to false. Now it's possible to add a little logic to set the modified to true. Which means all the neo fields for the entry are also getting duplicated.